Got rid of a bunch of gcc warnings and some potentially serious issues due to uninitialized variables.
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@957 01821946-851b-0410-b7f4-f1477c70e57a

+++ main/cameras.c
@@ -46,7 +46,7 @@ if (!OglCreateFBuffer (&pc->fb, pc->texBuf.bmProps.w, pc->texBuf.bmProps.h, pc->
 	return 0;
 pc->glTexId = pc->fb.hRenderBuffer;
 #endif
-sprintf (pc->texBuf.szName, "CAM#%04d", CAM_IDX (pc));
+sprintf (pc->texBuf.szName, "CAM#%04d", (int) CAM_IDX (pc));
 return 1;
 }
