+++ main/perlin.c
@@ -3,11 +3,11 @@
 
 #include "perlin.h"
 
-static int randSeed = 0;
+static long randSeed = 0;
 
 //------------------------------------------------------------------------------
 
-double Noise1D (int x)			 
+double Noise1D (long x)			 
 {
 #if 1
 srand (randSeed * (x + 2));
@@ -16,9 +16,9 @@ return (double) (x - ((RAND_MAX + 1) / 2)) / (double) ((RAND_MAX + 1) / 2);
 #else
 double h;
 x = randSeed + x * x;
-h = (int) pow ((double) x * (double) (1 << 13), x);
+h = (long) pow ((double) x * (double) (1 << 13), x);
 h = h * (h * (h * 15731 + 789221) + 1376312589);
-return 1.0 - ((int) h & 0x7fffffff) / 1073741824.0;    
+return 1.0 - ((long) h & 0x7fffffff) / 1073741824.0;    
 #endif
 }
 
@@ -57,7 +57,7 @@ return bObjectRendered;
 
 //------------------------------------------------------------------------------
 
-double SmoothedNoise1D (int x)
+double SmoothedNoise1D (long x)
 {
 return Noise1D (x) / 2  +  Noise1D (x-1) / 4  +  Noise1D (x+1) / 4;
 }
@@ -67,17 +67,17 @@ return Noise1D (x) / 2  +  Noise1D (x-1) / 4  +  Noise1D (x+1) / 4;
 double InterpolatedNoise1D (double x)
 {
 double xInt, xFrac = modf (x, &xInt);
-double v1 = SmoothedNoise1D ((int) xInt);
-double v2 = SmoothedNoise1D ((int) xInt + 1);
+double v1 = SmoothedNoise1D ((long) xInt);
+double v2 = SmoothedNoise1D ((long) xInt + 1);
 return CosineInterpolate (v1, v2, xFrac);
 }
 
 //------------------------------------------------------------------------------
 
-double PerlinNoise1D (double x, double persistence, int octaves, int seed)
+double PerlinNoise1D (double x, double persistence, long octaves, long seed)
 {
 double total = 0, frequency = 1.0, amplitude = 1.0;
-int i;
+long i;
 randSeed = seed;
 for (i = 0; i < octaves; i++) {
 	total += InterpolatedNoise1D (x * frequency) * amplitude;
@@ -89,14 +89,14 @@ return total;
 
 //------------------------------------------------------------------------------
 
-double Noise2D (int x, int y)
+double Noise2D (long x, long y)
 {
 return Noise1D (x + y * 57);
 }
 
 //------------------------------------------------------------------------------
 
-double SmoothedNoise2D (int x, int y)
+double SmoothedNoise2D (long x, long y)
 {
 double corners = (Noise2D (x-1, y-1) + Noise2D (x+1, y-1) + Noise2D (x-1, y+1) + Noise2D (x+1, y+1)) / 16;
 double sides = (Noise2D (x-1, y) + Noise2D (x+1, y) + Noise2D (x, y-1) + Noise2D (x, y+1)) / 8;
@@ -111,10 +111,10 @@ double InterpolatedNoise2D (double x, double y)
 double xInt, yInt,
 		 xFrac = modf (x, &xInt),
 		 yFrac = modf (x, &yInt);
-double v1 = SmoothedNoise2D ((int) xInt, (int) yInt);
-double v2 = SmoothedNoise2D ((int) xInt+1, (int) yInt);
-double v3 = SmoothedNoise2D ((int) xInt, (int) yInt+1);
-double v4 = SmoothedNoise2D ((int) xInt+1, (int) yInt+1);
+double v1 = SmoothedNoise2D ((long) xInt, (long) yInt);
+double v2 = SmoothedNoise2D ((long) xInt+1, (long) yInt);
+double v3 = SmoothedNoise2D ((long) xInt, (long) yInt+1);
+double v4 = SmoothedNoise2D ((long) xInt+1, (long) yInt+1);
 double i1 = CosineInterpolate (v1, v2, xFrac);
 double i2 = CosineInterpolate (v3, v4, xFrac);
 return CosineInterpolate (i1, i2, yFrac);
@@ -122,10 +122,10 @@ return CosineInterpolate (i1, i2, yFrac);
 
 //------------------------------------------------------------------------------
 
-double PerlinNoise2D (double x, double y, double persistence, int octaves, int nSeed)
+double PerlinNoise2D (double x, double y, double persistence, long octaves, long nSeed)
 {
 double total = 0, frequency = 1.0, amplitude = 1.0;
-int i;
+long i;
 randSeed = nSeed;
 for (i = 0; i < octaves; i++) {
 	total += InterpolatedNoise1D (x * frequency) * amplitude;
